requirements.txt
click
python-dotenv
snowflake-connector-python
tabulate